+++ src/system/config_client.rs
@@ -163,6 +163,23 @@ impl ConfigClient {
         };
     }
 
+    /// ### get_file_fmt
+    ///
+    /// Get current file fmt
+    pub fn get_file_fmt(&self) -> Option<String> {
+        self.config.user_interface.file_fmt.clone()
+    }
+
+    /// ### set_file_fmt
+    ///
+    /// Set file fmt parameter
+    pub fn set_file_fmt(&mut self, s: String) {
+        self.config.user_interface.file_fmt = match s.is_empty() {
+            true => None,
+            false => Some(s),
+        };
+    }
+
     // SSH Keys
 
     /// ### save_ssh_key
@@ -449,6 +466,21 @@ mod tests {
         assert_eq!(client.get_group_dirs(), Some(GroupDirs::First),);
         client.set_group_dirs(None);
         assert_eq!(client.get_group_dirs(), None,);
+    }
+
+    #[test]
+    fn test_system_config_file_fmt() {
+        let tmp_dir: tempfile::TempDir = create_tmp_dir();
+        let (cfg_path, key_path): (PathBuf, PathBuf) = get_paths(tmp_dir.path());
+        let mut client: ConfigClient = ConfigClient::new(cfg_path.as_path(), key_path.as_path())
+            .ok()
+            .unwrap();
+        assert_eq!(client.get_file_fmt(), None);
+        client.set_file_fmt(String::from("{NAME}"));
+        assert_eq!(client.get_file_fmt().unwrap(), String::from("{NAME}"));
+        // Delete
+        client.set_file_fmt(String::from(""));
+        assert_eq!(client.get_file_fmt(), None);
     }
 
     #[test]